Fix: 검색어 빈값일때 dataList나오는문제 해결
