+++ src/pages/SearchPage.js
@@ -3,37 +3,25 @@ import styled from 'styled-components';
 import { useGetSearchWordQuery } from '../services/searchWord';
 import { useDispatch } from 'react-redux';
 import { dataListUpdate } from '../services/dataListSlice';
+import useDebounce from '../util/useDebounce';
 
 import { SearchInput } from '../components/SearchInput';
 import { SearchList } from '../components/SearchList';
 import { Title } from '../components/Title';
 
-function useDebounce(value, delay) {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-}
-
 function SearchPage() {
 
 	const [inputValue, setInputValue] = useState('');
-	const debounceInputValue = useDebounce(inputValue, 1000);
+	const [isSearching, setIsSearching] = useState(false);
+	const debounceInputValue = useDebounce(inputValue, 700);
 	const [dataList, setDataList] = useState();
 	const [isFocus, setIsFocus] = useState(false);
 	const [selected, setSelected] = useState();
 	const [listIndex, setListIndex] = useState(-1);
-  const dispatch = useDispatch();
-  const { data } = useGetSearchWordQuery(debounceInputValue);
+	const dispatch = useDispatch();
+	const { data } = useGetSearchWordQuery(debounceInputValue, 
+		{ skip: debounceInputValue === '' ? true : false }
+	);
 
 	const handleKeyPress = (e) => {
 		if(e.keyCode === 40 && listIndex < 6){
@@ -45,6 +33,11 @@ function SearchPage() {
 		}
 	}
 
+	const handleInitSelelcted = () => {
+		setListIndex(-1);
+		setSelected();
+	}
+
 	useEffect(() => {
 		if(dataList && listIndex !== -1){
       setSelected(dataList[listIndex].id);
@@ -54,8 +47,7 @@ function SearchPage() {
 	//포커싱 아웃 되었을때 인덱스 초기화
 	useEffect(() => {
 		if(!isFocus) {
-			setListIndex(-1);
-      setSelected();
+			handleInitSelelcted();
 		}
 	}, [isFocus])
 
@@ -64,11 +56,17 @@ function SearchPage() {
     dispatch(dataListUpdate(data))
 	}, [data])
 
+	useEffect(() => {
+		handleInitSelelcted();
+		setIsSearching(false)
+	}, [debounceInputValue])
+
 	const handleInputValue = (e) => {
     setInputValue(e.target.value)
   }
 	const handleSearch = () => {
     setInputValue('')
+		setIsSearching(true)
 	}
 
 	return (
@@ -83,7 +81,8 @@ function SearchPage() {
           handleKeyPress={handleKeyPress}
         />
       </InputContainer>
-			<SearchList 
+			<SearchList
+				isSearching={isSearching}
 				isFocus={isFocus} 
 				selected={selected} 
 				inputLength={inputValue.length}
